+++ calendar/src/main/kotlin/ui/styles/NoteStyles.kt
@@ -1,4 +1,4 @@
-package frame.styles
+package ui.styles
 
 import javafx.scene.paint.Color
 import tornadofx.Stylesheet
