+++ calendar/src/main/kotlin/ui/styles/WeekStyles.kt
@@ -1,6 +1,6 @@
-package frame.styles
+package ui.styles
 
-import frame.styles.GlobalStyles.Mixins.BGColor
+import ui.styles.GlobalStyles.Mixins.BGColor
 import javafx.scene.paint.Color
 import tornadofx.*
